TaskController relate to auth user

+++ app/Controllers/Http/TaskController.js
@@ -6,13 +6,13 @@ const Task = use('App/Models/Task')
 const { validate } = use('Validator')
 
 class TaskController {
-    async index({ view }) {
-        const tasks = await Task.all()
+    async index({ view, auth }) {
+        const tasks = await auth.user.tasks().fetch()
 
         return view.render('tasks.index', { tasks: tasks.toJSON() })
     }
 
-    async store({ request, response, session }) {
+    async store({ request, auth, session, response }) {
         const validation = await validate(request.all(), {
             title: 'required|min:3|max:255',
             detail: 'max:512'
@@ -27,6 +27,7 @@ class TaskController {
         const task = new Task()
         task.title = request.input('title')
         task.detail = request.input('detail')
+        task.user_id = auth.user.id
         await task.save()
 
         session.flash({
@@ -36,14 +37,19 @@ class TaskController {
         return response.redirect('back')
     }
 
-    async destroy( {params, session, response} ) {
+    async destroy( {params, auth, session, response} ) {
         const task = await Task.find(params.id)
+
+        if (auth.user.id !== task.user_id) {
+            return response.redirect('/tasks')
+        }
+
         const title = task.title
         await task.delete()
       
         session.flash({ notification: `Task "${title}" deleted!` })
       
-        return response.redirect('back')
+        return response.redirect('/tasks')
     }
 }
